requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
python-dotenv
psycopg[binary]
matplotlib
catboost
lightgbm
